refactor(pixtuoid): split install-WRITE helpers out of verify.rs into merge.rs
verify.rs documented itself as the READ-ONLY #309 soundness detector but also
hosted the install-WRITE shared helpers (config parse, the sentinel-keyed hook
merge). Move them to a new sibling `install/merge.rs` (byte-identical bodies):
parse_json_or_empty, parse_toml_or_empty, bake_hook_path, hook_path_str,
flat_json_merge_install, flat_json_merge_uninstall, is_flat_managed. verify.rs
keeps only the read side (ShimRef/SchemaParse/SchemaVerifyResult, display_safe,
assemble, flat_json_verify, shell_shim_ref, posix_unquote).

The ~29 call sites across the 7 JSON/TOML targets (claude, codex, cursor,
reasonix, opencode, openclaw, codewhale) switch to `crate::install::merge::`.
The write-side unit tests move with their fns; verify.rs's module-doc header
now points at the merge sibling and the pixtuoid CLAUDE.md gains a merge.rs
bullet. Pure move + re-path — the per-target install round-trip tests cover it.
