+++ crates/pixtuoid/src/install/codewhale.rs
@@ -184,12 +184,12 @@ pub fn detect_installed() -> bool {
 pub fn hook_command(resolved: &Path, _explicit: bool) -> Result<String> {
     // `_explicit` is Claude's bare-name-vs-absolute switch — CodeWhale always
     // embeds the absolute path, so the flag changes nothing here.
-    let p = crate::install::verify::hook_path_str(resolved)?;
+    let p = crate::install::merge::hook_path_str(resolved)?;
     crate::install::hook_cmd::shell_hook_command(p, "codewhale")
 }
 
 pub fn merge_install(content: &str, base_cmd: &str) -> Result<MergeOutcome> {
-    let doc = crate::install::verify::parse_toml_or_empty(content)?;
+    let doc = crate::install::merge::parse_toml_or_empty(content)?;
     let merged = toml_merge_install(doc.clone(), base_cmd);
     let changed = merged != doc;
     Ok(MergeOutcome {
@@ -199,7 +199,7 @@ pub fn merge_install(content: &str, base_cmd: &str) -> Result<MergeOutcome> {
 }
 
 pub fn merge_uninstall(content: &str) -> Result<MergeOutcome> {
-    let doc = crate::install::verify::parse_toml_or_empty(content)?;
+    let doc = crate::install::merge::parse_toml_or_empty(content)?;
     let cleaned = toml_merge_uninstall(doc.clone());
     let changed = cleaned != doc;
     Ok(MergeOutcome {
