+++ crates/pixtuoid/src/install/openclaw.rs
@@ -216,7 +216,7 @@ fn resolve_openclaw_detect(
 /// The shim's absolute path, baked into the plugin (the gateway runs it under
 /// Node, no PATH reliance). Err on non-UTF-8 like opencode/Codex.
 pub fn hook_command(resolved: &Path, _explicit: bool) -> Result<String> {
-    crate::install::verify::hook_path_str(resolved).map(str::to_string)
+    crate::install::merge::hook_path_str(resolved).map(str::to_string)
 }
 
 /// The wholly-owned plugin dir files (manifest + package.json + entry module).
@@ -234,7 +234,7 @@ pub fn plugin_artifacts(hook_path: &Path) -> Result<Vec<(PathBuf, String)>> {
 }
 
 fn render_plugin(hook_path: &str) -> Result<String> {
-    crate::install::verify::bake_hook_path(PLUGIN_TEMPLATE, HOOK_PLACEHOLDER, hook_path, "openclaw")
+    crate::install::merge::bake_hook_path(PLUGIN_TEMPLATE, HOOK_PLACEHOLDER, hook_path, "openclaw")
 }
 
 fn obj_mut<'a>(v: &'a mut Value, key: &str) -> Result<&'a mut serde_json::Map<String, Value>> {
@@ -256,7 +256,7 @@ pub fn merge_install(content: &str, _hook_cmd: &str) -> Result<MergeOutcome> {
         .ok_or_else(|| anyhow!("plugin dir path is non-UTF-8: {}", dir.display()))?
         .to_string();
     let mut root =
-        crate::install::verify::parse_json_or_empty(content).context("parsing openclaw.json")?;
+        crate::install::merge::parse_json_or_empty(content).context("parsing openclaw.json")?;
     let before = root.clone();
     {
         let root_obj = obj_mut(&mut root, "root")?;
@@ -294,7 +294,7 @@ pub fn merge_uninstall(content: &str) -> Result<MergeOutcome> {
     let dir = plugin_dir()?;
     let dir_str = dir.to_str().map(str::to_string);
     let mut root =
-        crate::install::verify::parse_json_or_empty(content).context("parsing openclaw.json")?;
+        crate::install::merge::parse_json_or_empty(content).context("parsing openclaw.json")?;
     let before = root.clone();
     if let Some(plugins) = root.get_mut("plugins").and_then(Value::as_object_mut) {
         if let Some(paths) = plugins
